Major Refactoring
